+++ game_files/constants.py
@@ -1,0 +1,10 @@
+import pygame
+
+FRAMERATE = 60
+AIM_INDICATOR_DISTANCE_FROM_PLAYER = 50
+OFFSET_OF_GUN_FROM_PLAYER = pygame.Vector2(0,0)
+WEAPON_CAP_AMOUNT = 6
+WEAPON_SPAWN_RATE_IN_MILLIS = 10_000
+PLAYER_ACCELERATION_RATE = 0.5
+PLAYER_FRICTION = -0.5
+PLAYER_MAX_VELOCITY = 6
