--- tempCodeRunnerFile.py
@@ -1,2 +1,0 @@
-
-pygame.init()
